Allow automatic updates of ValueSuggest values (IdRef only)
Some suggesters returns a label and a URI. They are both stored in the
database.  But it's possible that the label associated to that URI
change over time, making the label in Omeka outdated.

This patch adds the ability for data types to declare an "updater" that
will take care of retrieving up-to-date informations and updating the
value accordingly.
This update process is accessible through a new item in the admin
navigation menu.  It is possible to limit which data types should be
checked and updated.
The update will be done in a background job.

This patch implements an updater for IdRef data types.

Test plan:
1. Create a resource template with dcterms:subject linked to data type
   "IdRef: Subjects"
2. Create an item with this resource template, and add a subject:
   - Start typing (for instance, "rome" will return several suggestions)
   - Select one of the suggestions
3. Once the suggester has filled the URI field, modify slightly the
   label (add or remove some characters) and save the item.
4. Click on "Value Suggest" in the admin sidebar
5. Check "IdRef: Subjects" checkbox and submit the form
6. Go to "Jobs", wait until the job is complete (it should not take more
   than a second)
7. Verify in your item that the value was correctly updated to its
   original value.